Moved MainActivity to ui package and Fragments to ui.fragments package
