+++ app/src/main/java/com/saotome/imed_app/ui/fragments/SecondFragment.kt
@@ -1,4 +1,4 @@
-package com.saotome.imed_app
+package com.saotome.imed_app.ui.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,6 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
+import com.saotome.imed_app.R
 import com.saotome.imed_app.databinding.FragmentSecondBinding
 
 /**
