+++ app/src/main/res/layout/fragment_drug_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DrugListFragment">
+    tools:context=".ui.fragments.DrugListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_drugs"
