+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="com.saotome.imed_app.DrugListFragment"
+        android:name="com.saotome.imed_app.ui.fragments.DrugListFragment"
         android:label="@string/drug_list_fragment_label"
         tools:layout="@layout/fragment_drug_list">
 
@@ -17,7 +17,7 @@
     </fragment>
     <fragment
         android:id="@+id/SecondFragment"
-        android:name="com.saotome.imed_app.SecondFragment"
+        android:name="com.saotome.imed_app.ui.fragments.SecondFragment"
         android:label="@string/second_fragment_label"
         tools:layout="@layout/fragment_second">
