+++ app/src/main/res/values/strings.xml
@@ -2,7 +2,7 @@
     <string name="app_name">imed-app</string>
     <string name="action_settings">Settings</string>
     <!-- Strings used for fragments for navigation -->
-    <string name="first_fragment_label">Drugs</string>
+    <string name="drug_list_fragment_label">Drugs</string>
     <string name="second_fragment_label">Second Fragment</string>
     <string name="next">Next</string>
     <string name="previous">Previous</string>
